Change init & exit functions to register multiple devices
Following Chapter 3 of LDD, we're only supporting scull0-3, so we can have a fixed amount of devices.

As we support other devices later on, we'll need to revise this. It might be beneficial to have a 'smart' way
of basing total device numbers on how many instances of each device type we have, but I'll leave this until we
have more than one type of device.